Add `PointRobot` env

+++ gymnax/environments/misc/gaussian_bandit.py
@@ -142,7 +142,7 @@ class GaussianBandit(environment.Environment):
 
 
 def time_normalization(
-    t: float, min_lim: float = -1.0, max_lim: float = 1.0, t_max: int = 100
+    t: int, min_lim: float = -1.0, max_lim: float = 1.0, t_max: int = 100
 ) -> float:
     """Normalize time integer into range given max time."""
     return (max_lim - min_lim) * t / t_max + min_lim
